+++ js/packages/cli/src/candy-machine-cli.ts
@@ -459,6 +459,8 @@ programCommand('show')
       log.info('price: ', machine.data.price.toNumber());
       //@ts-ignore
       log.info('itemsAvailable: ', machine.data.itemsAvailable.toNumber());
+      //@ts-ignore
+      log.info('itemsRedeemed: ', machine.itemsRedeemed.toNumber());
       log.info(
         'goLiveDate: ',
         //@ts-ignore
@@ -636,37 +638,57 @@ programCommand('update_candy_machine')
     '-r, --rpc-url <string>',
     'custom rpc url since this is a heavy command',
   )
+  .option('--new-authority <Pubkey>', 'New Authority. Base58-encoded')
   .action(async (directory, cmd) => {
-    const { keypair, env, date, rpcUrl, price, cacheName } = cmd.opts();
+    const { keypair, env, date, rpcUrl, price, newAuthority, cacheName } =
+      cmd.opts();
     const cacheContent = loadCache(cacheName, env);
 
     const secondsSinceEpoch = date ? parseDate(date) : null;
     const lamports = price ? parsePrice(price) : null;
+    const newAuthorityKey = newAuthority ? new PublicKey(newAuthority) : null;
 
     const walletKeyPair = loadWalletKey(keypair);
     const anchorProgram = await loadCandyProgram(walletKeyPair, env, rpcUrl);
 
     const candyMachine = new PublicKey(cacheContent.candyMachineAddress);
-    const tx = await anchorProgram.rpc.updateCandyMachine(
-      lamports ? new anchor.BN(lamports) : null,
-      secondsSinceEpoch ? new anchor.BN(secondsSinceEpoch) : null,
-      {
+
+    if (lamports || secondsSinceEpoch) {
+      const tx = await anchorProgram.rpc.updateCandyMachine(
+        lamports ? new anchor.BN(lamports) : null,
+        secondsSinceEpoch ? new anchor.BN(secondsSinceEpoch) : null,
+        {
+          accounts: {
+            candyMachine,
+            authority: walletKeyPair.publicKey,
+          },
+        },
+      );
+
+      cacheContent.startDate = secondsSinceEpoch;
+      if (date)
+        log.info(
+          ` - updated startDate timestamp: ${secondsSinceEpoch} (${date})`,
+        );
+      if (lamports)
+        log.info(` - updated price: ${lamports} lamports (${price} SOL)`);
+      log.info('update_candy_machine finished', tx);
+    }
+
+    if (newAuthorityKey) {
+      const tx = await anchorProgram.rpc.updateAuthority(newAuthorityKey, {
         accounts: {
           candyMachine,
           authority: walletKeyPair.publicKey,
         },
-      },
-    );
+      });
 
-    cacheContent.startDate = secondsSinceEpoch;
+      cacheContent.authority = newAuthorityKey.toBase58();
+      log.info(` - updated authority: ${newAuthorityKey.toBase58()}`);
+      log.info('update_authority finished', tx);
+    }
+
     saveCache(cacheName, env, cacheContent);
-    if (date)
-      log.info(
-        ` - updated startDate timestamp: ${secondsSinceEpoch} (${date})`,
-      );
-    if (lamports)
-      log.info(` - updated price: ${lamports} lamports (${price} SOL)`);
-    log.info('update_candy_machine finished', tx);
   });
 
 programCommand('mint_one_token')
@@ -679,7 +701,13 @@ programCommand('mint_one_token')
 
     const cacheContent = loadCache(cacheName, env);
     const configAddress = new PublicKey(cacheContent.program.config);
-    const tx = await mint(keypair, env, configAddress, rpcUrl);
+    const tx = await mint(
+      keypair,
+      env,
+      configAddress,
+      cacheContent.program.uuid,
+      rpcUrl,
+    );
 
     log.info('mint_one_token finished', tx);
   });