+++ js/packages/graphql/src/common/models/auctions/decode/decodeBidderPot.ts
@@ -1,5 +1,0 @@
-import { decodeEntity } from '../../BaseEntity';
-import { BidderPot } from '../entities';
-import { AUCTION_SCHEMA } from '../schema';
-
-export const decodeBidderPot = decodeEntity(BidderPot, AUCTION_SCHEMA);