+++ js/packages/graphql/src/common/models/metadata/decode/index.ts
@@ -1,3 +1,0 @@
-export * from './decodeEdition';
-export * from './decodeMasterEdition';
-export * from './decodeMetadata';