+++ js/packages/graphql/src/common/models/metaplex/BidRedemptionTicket.ts
@@ -1,7 +1,0 @@
-import type { MetaplexKey } from './MetaplexKey';
-
-export interface BidRedemptionTicket {
-  key: MetaplexKey;
-
-  getBidRedeemed(order: number): boolean;
-}