+++ js/packages/graphql/src/common/models/metaplex/ClaimBidArgs.ts
@@ -1,3 +1,0 @@
-export class ClaimBidArgs {
-  instruction = 6;
-}