+++ js/packages/graphql/src/common/models/metaplex/DeprecatedValidateParticipationArgs.ts
@@ -1,3 +1,0 @@
-export class DeprecatedValidateParticipationArgs {
-  instruction = 10;
-}