+++ js/packages/graphql/src/common/models/metaplex/WinningConstraint.ts
@@ -1,4 +1,0 @@
-export enum WinningConstraint {
-  NoParticipationPrize = 0,
-  ParticipationPrizeGiven = 1,
-}