+++ js/packages/graphql/src/common/models/metaplex/decode/decodeSafetyDepositConfig.ts
@@ -1,8 +1,0 @@
-import { SafetyDepositConfig } from '../entities';
-
-export function decodeSafetyDepositConfig(buffer: Buffer, pubkey?: string) {
-  return new SafetyDepositConfig({
-    data: buffer,
-    pubkey,
-  });
-}