+++ js/packages/web/src/components/ArtCard/index.tsx
@@ -28,6 +28,8 @@ export interface ArtCardProps extends CardProps {
   height?: number;
   artView?: boolean;
   width?: number;
+
+  count?: string;
 }
 
 export const ArtCard = (props: ArtCardProps) => {
@@ -46,6 +48,7 @@ export const ArtCard = (props: ArtCardProps) => {
     height,
     artView,
     width,
+    count,
     ...rest
   } = props;
   const art = useArt(pubkey);
@@ -111,6 +114,9 @@ export const ArtCard = (props: ArtCardProps) => {
               </>
             )} */}
             <div className="edition-badge">{badge}</div>
+            {count && (
+              <div className="edition-badge">Selected count: {count}</div>
+            )}
           </>
         }
       />