+++ js/packages/web/src/components/AuctionCard/hooks/useInstantSaleState.ts
@@ -3,6 +3,7 @@ import { AuctionView } from '../../../hooks';
 
 interface ActionButtonContentProps {
   isInstantSale: boolean;
+  isAlreadyBought: boolean;
   canClaimItem: boolean;
   canClaimPurchasedItem: boolean;
   canEndInstantSale: boolean;
@@ -13,16 +14,30 @@ export const useInstantSaleState = (
 ): ActionButtonContentProps => {
   const wallet = useWallet();
 
-  const { isInstantSale, auctionManager, auction, myBidderPot } = auctionView;
+  const { isInstantSale, auctionManager, auction, myBidRedemption } =
+    auctionView;
 
+  const items = auctionView.items;
+  let isAlreadyBought: boolean = false;
+
+  for (const item of items) {
+    for (const subItem of item) {
+      const bidRedeemed = myBidRedemption?.info.getBidRedeemed(
+        subItem.safetyDeposit.info.order,
+      );
+      isAlreadyBought = bidRedeemed ? bidRedeemed : false;
+      if (isAlreadyBought) break;
+    }
+  }
+  const canClaimPurchasedItem = isAlreadyBought;
   const isOwner = auctionManager.authority === wallet?.publicKey?.toBase58();
   const isAuctionEnded = auction.info.endedAt;
-  const canClaimPurchasedItem = !!myBidderPot;
   const canClaimItem = !!(isOwner && isAuctionEnded);
   const canEndInstantSale = isOwner && !isAuctionEnded;
 
   return {
     isInstantSale,
+    isAlreadyBought,
     canClaimItem,
     canClaimPurchasedItem,
     canEndInstantSale,