+++ js/packages/web/src/components/AuctionRenderCard/index.tsx
@@ -1,6 +1,5 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Card, CardProps } from 'antd';
-import { BidStateType, CountdownState } from '@oyster/common';
 import { ArtContent } from '../ArtContent';
 import { AuctionView, useArt, useCreators } from '../../hooks';
 import { AmountLabel } from '../AmountLabel';
@@ -8,6 +7,7 @@ import { MetaAvatar } from '../MetaAvatar';
 import { AuctionCountdown } from '../AuctionNumbers';
 
 import { useAuctionStatus } from './hooks/useAuctionStatus';
+import { useTokenList } from '../../contexts/tokenList';
 
 export interface AuctionCard extends CardProps {
   auctionView: AuctionView;
@@ -20,6 +20,7 @@ export const AuctionRenderCard = (props: AuctionCard) => {
   const creators = useCreators(auctionView);
   const name = art?.title || ' ';
 
+  const tokenInfo = useTokenList().mainnetTokens.filter(m=>m.address == auctionView.auction.info.tokenMint)[0]
   const { status, amount } = useAuctionStatus(auctionView);
 
   const card = (
@@ -55,6 +56,7 @@ export const AuctionRenderCard = (props: AuctionCard) => {
           title={status}
           amount={amount}
           iconSize={24}
+          tokenInfo={tokenInfo}
         />
       </div>
     </Card>