+++ js/packages/web/src/views/packCreate/components/DesignAndInfoStep/interface.ts
@@ -1,0 +1,7 @@
+import { FormInstance } from 'antd';
+import { InfoFormState, PackState } from '../../interface';
+
+export interface DesignAndInfoStepProps {
+  setPackState: (values: Partial<PackState>) => void;
+  form: FormInstance<InfoFormState>;
+}